+++ test/test_vlasov_solvers.jl
@@ -7,4 +7,5 @@ end
     include(joinpath("VlasovSolver", "test_nonuniform_advection.jl"))
     include(joinpath("VlasovSolver", "test_strang_splitting.jl"))
     include(joinpath("VlasovSolver", "test_free_streaming.jl"))
+    include(joinpath("VlasovSolver", "test_echo.jl"))
 end